move sdxl controlnet export
